pyproject.toml
[project]
name = "aav-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dotenv>=0.9.9",
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "google-auth-oauthlib>=1.3.1",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "lxml>=6.0.2",
    "openai>=2.30.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
